Organizando arquivos e adicionando comentários
