--- src/timeline/export-firestore.js
@@ -1,40 +1,0 @@
-import {
-  collection,
-  addDoc,
-  getFirestore,
-  getDocs,
-  orderBy,
-  query,
-  // doc,
-  // deleteDoc,
-} from '../firebase/firestore.js';
-
-const db = getFirestore();
-
-export async function addPosts(message, userEmail) {
-  try {
-    const docRef = await addDoc(collection(db, 'posts'), {
-      message,
-      userEmail,
-      date: new Date().toLocaleString('pt-br'),
-    });
-    console.log('Document written with ID: ', docRef.id);
-    return docRef.id;
-  } catch (e) {
-    console.error('Error adding document: ', e);
-    return null;
-  }
-}
-
-export const getPosts = async () => {
-  const arrPosts = [];
-  const orderFirestore = query(collection(db, 'posts'), orderBy('date'));
-  const querySnapshot = await getDocs(orderFirestore);
-  querySnapshot.forEach((item) => {
-    const timeline = item.data();
-    timeline.id = item.id;
-    arrPosts.push(timeline);
-  });
-
-  return arrPosts;
-};
